refactor(consensus): move EpochState to state.rs
