+++ core/consensus/src/lib.rs
@@ -2,6 +2,7 @@ pub mod config;
 pub mod consensus;
 pub mod execution;
 pub mod narwhal;
+mod state;
 #[cfg(test)]
 mod tests;
 pub mod validator;
